projects/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
fastapi==0.72.0
googletrans==3.0.0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
MarkupSafe==2.0.1
MouseInfo==0.1.3
mysql-connector-python==8.0.30
nltk==3.7
passlib==1.7.4
Pillow==9.0.0
protobuf==3.20.1
psycopg2==2.9.3
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.9.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dotenv==0.19.2
pytweening==1.0.4
pywhatkit==5.3
PyYAML==5.4.1
regex==2022.6.2
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.35
starlette==0.17.1
textblob==0.17.1
tqdm==4.64.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
wikipedia==1.4.0
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
